implemented red test BalancingServerWithOneSlotCapacityAndOneSlotVMFillsServerWithVM

+++ LoadBalancerMTO/Server.cs
@@ -7,6 +7,6 @@ namespace LoadBalancerMTO
     public class Server
     {
         public double CurrentLoadPercentage = 0d;
-        public double Capacity = 0d;
+        public int Capacity = 0;
     }
 }
